pyproject.toml
[project]
name = "angel-smart-trader"
version = "0.1.0"
description = "Server-side Angel One SmartAPI trading pipeline (auth → data → strategy → risk → execution)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "httpx>=0.27,<0.29",
  "pydantic>=2.6,<3",
  "pydantic-settings>=2.2,<3",
  "pyotp>=2.9,<3",
  "structlog>=24.1,<26",
  "uvicorn[standard]>=0.30,<1",
  "websocket-client>=1.7,<2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.6"]

[project.scripts]
angel-trader = "angel_bot.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/angel_bot"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
